chore: final implementation
