--- packages/cloud-ide/config-overrides.js
@@ -1,19 +1,0 @@
-
-module.exports = function override(config) {
-    const fallback = config.resolve.fallback || {};
-    Object.assign(fallback, {
-        path: require.resolve("path-browserify"),
-        process: false,
-    });
-    config.resolve.fallback = fallback;
-    config.ignoreWarnings = [/Failed to parse source map/];
-    config.module.rules.push({
-        test: /\.(ts|mjs|tsx)$/,
-        enforce: "pre",
-        loader: require.resolve("source-map-loader"),
-        resolve: {
-            fullySpecified: false,
-        },
-    });
-    return config;
-};
