+++ packages/cloud-ide/package.json
@@ -5,21 +5,18 @@
   "dependencies": {
     "@monaco-editor/react": "^4.5.2",
     "@quik/ui": "workspace:*",
-    "@testing-library/jest-dom": "^5.17.0",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
     "@webcontainer/api": "^1.1.5",
     "dockview": "^1.8.3",
     "monaco-editor": "^0.39.0",
     "monaco-editor-auto-typings": "^0.4.4",
     "react": "^18.2.0",
-    "react-app-rewired": "^2.2.1",
     "react-complex-tree": "^2.2.2",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.16.0",
     "react-scripts": "5.0.1",
     "styled-components": "^6.0.8",
     "typescript": "^4.9.5",
+    "web-vitals": "^2.1.4",
     "xterm": "^5.3.0",
     "xterm-addon-fit": "^0.7.0",
     "y-monaco": "^0.1.4",
@@ -27,18 +24,19 @@
     "yjs": "^13.6.8"
   },
   "devDependencies": {
-    "@types/jest": "^27.5.2",
-    "@types/node": "^16.18.54",
+    "@types/node": "^20.7.0",
     "@types/react": "^18.2.23",
     "@types/react-dom": "^18.2.8",
-    "@types/react-router-dom": "^5.3.3",
-    "@types/styled-components": "^5.1.28",
-    "path-browserify": "^1.0.1"
+    "@vitejs/plugin-react": "^4.1.0",
+    "@vitejs/plugin-react-swc": "^3.4.0",
+    "vite": "^4.4.9",
+    "vite-plugin-node-polyfills": "^0.9.0",
+    "vite-plugin-rewrite-all": "^1.0.1"
   },
   "scripts": {
-    "start": "react-app-rewired start",
-    "build": "react-app-rewired build",
-    "test": "react-app-rewired test",
+    "start": "vite",
+    "build": "tsc && vite build",
+    "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
