--- packages/cloud-ide/src/components/CodePlayground/index.tsx
@@ -1,8 +1,0 @@
-import React from "react";
-import FileExplorer from "./FileExplorer";
-
-const CodePlayground: React.FC = () => {
-    return <FileExplorer />;
-}
-
-export default CodePlayground;
