--- packages/cloud-ide/src/components/Terminal/index.tsx
@@ -1,29 +1,0 @@
-import { useRef, useEffect } from 'react';
-
-import type { GridviewPanelApi } from 'dockview';
-import type { ShellInstance, ServerReadyHandler } from '../../hooks/useShell';
-
-interface TerminalProps {
-    shell: ShellInstance,
-    panelApi: GridviewPanelApi,
-    onServerReady: ServerReadyHandler,
-}
-
-const Terminal = (props: TerminalProps) => {
-    const {shell} = props;
-    const init = useRef(false);
-    const root = useRef<HTMLDivElement>(null);
-
-    useEffect(() => {
-        if (!init.current && root.current) {
-            init.current = true;
-            shell.start(root.current, props.panelApi, props.onServerReady);
-        }
-    }, [shell]);
-
-    return (
-        <div ref={root} style={{height: '100%'}}></div>
-    );
-}
-
-export default Terminal;
