--- packages/cloud-ide/src/components/WaterMark/index.tsx
@@ -1,12 +1,0 @@
-import type { IWatermarkPanelProps } from 'dockview';
-
-const Watermark = (_props: IWatermarkPanelProps) => {
-    return (
-        <div className="watermark">
-            <div className="letterpress"/>
-            <div className="shortcuts"></div>
-        </div>
-    );
-}
-
-export default Watermark;
