+++ packages/cloud-ide/src/hooks/useCollab.ts
@@ -1,21 +1,22 @@
-import { useRef, useCallback, useEffect } from 'react';
+import {useRef, useCallback, useEffect} from 'react';
 
-import type { ShellInstance } from './useShell';
-import type { Editor } from '../libs/monaco';
-import type * as YJS from '../libs/yjs';
+
+import type {ShellInstance} from './useShell';
+import type {Editor} from '../modules/monaco';
+import type * as YJS from '../modules/yjs';
 
 export type CollabInstance = {
   session: React.MutableRefObject<YJS.Session | null>,
   syncEditor: (editor: Editor) => void,
 };
 
-const useCollab = (shell: ShellInstance): CollabInstance => {
+export function useCollab(shell: ShellInstance): CollabInstance {
   const init = useRef(false);
   const session = useRef<YJS.Session | null>(null);
   const syncMonaco = useRef<typeof YJS.monaco | null>(null);
   const syncEditor = useCallback((editor: Editor) => {
     if (!session.current || !syncMonaco.current) return;
-    console.log('Editor sync enabled.');
+    console.log('Index sync enabled.');
     syncMonaco.current(editor, session.current.document, session.current.provider);
   }, []);
 
@@ -23,11 +24,11 @@ const useCollab = (shell: ShellInstance): CollabInstance => {
     if (init.current) return;
     if (shell && shell.container?.fs) {
       init.current = true;
-      const syncKey = window.location.hash;
+      const syncKey = location.hash;
       if (!syncKey) return;
       console.log('Setting up sync using key:', syncKey);
       // Import sync module & connect
-      import('../libs/yjs').then(sync => {
+      import('../modules/yjs').then(sync => {
         session.current = sync.connect(syncKey);
         syncMonaco.current = sync.monaco;
         console.log('Connected to sync.');
@@ -35,7 +36,5 @@ const useCollab = (shell: ShellInstance): CollabInstance => {
     }
   }, [shell]);
 
-  return { session, syncEditor };
+  return {session, syncEditor};
 }
-
-export default useCollab;
