+++ packages/cloud-ide/src/index.css
@@ -1,3 +1,22 @@
-body {
-    margin: 0;
+:root {
+  text-rendering: optimizeLegibility;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  -webkit-text-size-adjust: 100%;
+}
+
+body, html, #root {
+  width: 100%;
+  height: 100%;
+  margin: 0;
+  overscroll-behavior: none;
+}
+
+.split-view-container .view-container {
+  background: #333;
+  color: #fff;
+}
+
+.xterm {
+  padding: 10px;
 }
