+++ packages/cloud-ide/src/modules/webcontainer.ts
@@ -3,7 +3,7 @@ import type {FileSystemAPI, FileSystemTree} from '@webcontainer/api';
 import type {TreeItem, TreeItemIndex} from 'react-complex-tree';
 import type {DockviewApi} from 'dockview';
 
-const configRaw = globalThis.localStorage?.vslite_config;
+const configRaw = globalThis.localStorage?.quiklabs_config;
 const config = configRaw ? JSON.parse(configRaw) : {};
 
 export async function getDirAsTree(
@@ -36,6 +36,8 @@ export async function getDirAsTree(
     if (parent) db?.[parent]?.children?.push(itemPath);
     if (isDir) await getDirAsTree(fs, itemPath, itemPath, root, db);
   };
+
+  console.log('getDirAsTree() db', db);
 
   return db;
 }
