+++ packages/cloud-ide/src/routes/AppRouter.tsx
@@ -1,17 +1,12 @@
-import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import Layout from '../layout';
-import Dock from "../components";
+import { Playground } from "../components/Playground";
 
 const AppRouter: React.FC = () => {
     return (
-        <Router>
-            <Layout>
-                <Routes>
-                    <Route path="/" element={<Dock/>} />
-                </Routes>
-            </Layout>
-        </Router>
+        <Layout>
+            <Playground/>
+        </Layout>
     );
 };
