+++ packages/cloud-ide/src/utils/debounce.ts
@@ -1,4 +1,3 @@
-
 export function debounce<T extends Function>(cb: T, wait = 150) {
   let h: NodeJS.Timeout;
   let callable = (...args: any) => {
