+++ packages/cloud-ide/src/vite-env.d.ts
@@ -1,0 +1,5 @@
+/// <reference types="vite/client" />
+
+declare module globalThis {
+  var process: Record<string, any>;
+}
