+++ packages/cloud-ide/tsconfig.json
@@ -1,6 +1,7 @@
 {
-    "extends": "../../tsconfig.json",
-    "include": [
-       "src"
-    ]
+  "include": ["src"],
+  "references": [{"path": "./tsconfig.node.json"}],
+  "compilerOptions": {
+    "jsx": "react-jsx"
+  }
 }
