+++ packages/cloud-ide/tsconfig.node.json
@@ -1,0 +1,11 @@
+{
+  "include": ["vite.config.ts"],
+  "compilerOptions": {
+    "target": "ES2022",
+    "module": "ESNext",
+    "moduleResolution": "node",
+    "allowSyntheticDefaultImports": true,
+    "skipLibCheck": true,
+    "composite": true
+  }
+}
