+++ packages/ui/package.json
@@ -10,7 +10,7 @@
         "build": "tsc --project ./tsconfig.json --outDir ./dist"
     },
     "devDependencies": {
-        "@types/react": "^18.2.22",
+        "@types/react": "^18.2.23",
         "react": "^18.2.0",
         "typescript": "^4.9.5"
     },
